Enable MySQL supprot

+++ osafw-app/App_Code/fw/FormUtils.cs
@@ -195,7 +195,7 @@ public class FormUtils
     }
 
     // return pager or Nothing if no paging required
-    public static ArrayList getPager(int count, int pagenum, object pagesize1 = null)
+    public static ArrayList getPager(long count, int pagenum, object pagesize1 = null)
     {
         int pagesize = MAX_PAGE_ITEMS;
         if (pagesize1 != null)
